# Speed up tests by removing code coverage and reduce number of runs on travis