--- a/.styleci.yml
+++ b/.styleci.yml
@@ -1,7 +1,0 @@
-preset: symfony
-
-enabled:
-    - align_equals
-
-disabled:
-    - phpdoc_annotation_without_dot # This is still buggy: https://github.com/symfony/symfony/pull/19198