--- a/.travis.yml
+++ b/.travis.yml
@@ -1,33 +1,17 @@
 language: php
 sudo: false
 
-php:
-    - 5.5
-    - 5.6
-    - 7.0
-
 env:
-    - BUILD_ALL=true
-    - ADAPTER=Apc
-    - ADAPTER=Apcu
-    - ADAPTER=Chain
-    - ADAPTER=Common
-    - ADAPTER=Doctrine
-    - ADAPTER=Filesystem
-    - ADAPTER=Memcache
-    - ADAPTER=Memcached
-    - ADAPTER=MongoDB
-    - ADAPTER=PHPArray
-    - ADAPTER=Predis
-    - ADAPTER=Redis
-    - ADAPTER=Void
-    - LIBRARY=Bridge
-    - LIBRARY=Hierarchy
-    - LIBRARY=Namespaced
-    - LIBRARY=SessionHandler
-    - LIBRARY=Taggable
+  global:
+    - COVERAGE=false
+    - TEST="./vendor/bin/phpunit"
 
 matrix:
+    include:
+      - php: 5.5
+      - php: 5.6
+      - php: 7.0
+        env: COVERAGE=true
     fast_finish: true
 
 services:
@@ -47,18 +31,16 @@ cache:
       - "$HOME/.composer/cache"
 
 before_install:
-    - echo "Disable xdebug" && phpenv config-rm xdebug.ini
-    - . ./build/setEnvVariables.sh
+    - if [[ $COVERAGE != true ]]; then phpenv config-rm xdebug.ini || true; fi
+    - if [[ $COVERAGE == true ]]; then TEST="$TEST --coverage-clover=coverage.xml"; fi
     - mkdir -p ~/.phpenv/versions/$(phpenv version-name)/etc/conf.d
-    - if [ -f "build/php/$TRAVIS_PHP_VERSION/$PROJECT.sh" ]; then . ./build/php/$TRAVIS_PHP_VERSION/$PROJECT.sh; fi
-    - if [ ! -z "$BUILD_ALL" ]; then . ./build/php/all.sh; fi
-    - travis_retry composer self-update
+    - ./build/php/all.sh
 
 script:
-    - ./build/runTest.sh "$DIR/$PROJECT"
+    - ./build/runAllTests.sh
 
 after_success:
-    - if [[ $TRAVIS_PHP_VERSION = 7.* ]]; then pip install --user codecov && codecov ; fi
+    - if [[ $COVERAGE  == true ]]; then pip install --user codecov && codecov ; fi
 
 notifications:
     email: false