--- a/build/runTest.sh
+++ b/build/runTest.sh
@@ -1,17 +1,21 @@
 #!/usr/bin/env bash
 
 cd $1
-SLUG=$(grep -Pho 'cache/[A-Za-z-\.]+' composer.json | xargs | awk '{print $1}')
+SLUG=$(grep -o 'cache/[A-Za-z\-\.]*' composer.json | xargs | awk '{print $1}')
 
-printf "\n\n************ Opening directory: $1 ************\n"
-printf     "************ Running tests for: $SLUG ************\n\n"
+printf "\n\n\n\n\n\n"
+printf "****************************************************************************************\n"
+printf "****************************************************************************************\n"
+printf "****************************************************************************************\n"
+printf "***\n"
+printf "***  Opening directory: $1 \n"
+printf "***  Running tests for: $SLUG \n"
+printf "***\n"
+printf "****************************************************************************************\n"
+printf "****************************************************************************************\n\n"
 
-if [ ! -z "$BUILD_ALL" ]; then composer require --no-update mongodb/mongodb:^1.0; fi
-if [ ! -z "$BUILD_ALL" ]; then composer require --no-update predis/predis:^1.0; fi
+if [ "$SLUG" = 'cache/cache' ]; then composer require --no-update mongodb/mongodb:^1.0 predis/predis:^1.0; fi
 
-composer install --no-interaction || exit -1
-
-TEST="./vendor/bin/phpunit"
-if [ ! -z "$BUILD_ALL" ]; then TEST="$TEST --coverage-clover=coverage.xml"; fi
+composer update --no-interaction || exit -1
 
 sh -c "$TEST" || exit -1