--- a/phpunit.xml.dist
+++ b/phpunit.xml.dist
@@ -22,6 +22,12 @@
     <log type="coverage-text" target="php://stdout"/>
   </logging>
 
+  <php>
+    <env name="MONGODB_HOST" value="localhost:27017"/>
+    <env name="MONGODB_COLLECTION" value="psr6test.cache"/>
+    <env name="MONGODB_DATABASE" value="test"/>
+  </php>
+
   <filter>
     <whitelist>
       <directory>./</directory>