--- a/src/Adapter/Memcache/phpunit.xml.dist
+++ b/src/Adapter/Memcache/phpunit.xml.dist
@@ -17,11 +17,6 @@
     </testsuite>
   </testsuites>
 
-  <logging>
-    <log type="coverage-text" target="php://stdout"/>
-    <log type="coverage-html" target="./coverageReport"/>
-  </logging>
-
   <groups>
     <exclude>
       <group>benchmark</group>