--- a/src/Adapter/Memcached/phpunit.xml.dist
+++ b/src/Adapter/Memcached/phpunit.xml.dist
@@ -16,12 +16,6 @@
       <directory>./Tests/</directory>
     </testsuite>
   </testsuites>
-
-  <logging>
-    <log type="coverage-text" target="php://stdout"/>
-    <log type="coverage-html" target="./coverageReport"/>
-  </logging>
-
   <groups>
     <exclude>
       <group>benchmark</group>