--- a/src/SessionHandler/phpunit.xml.dist
+++ b/src/SessionHandler/phpunit.xml.dist
@@ -17,10 +17,6 @@
     </testsuite>
   </testsuites>
 
-  <logging>
-    <log type="coverage-text" target="php://stdout"/>
-  </logging>
-
   <groups>
     <exclude>
       <group>benchmark</group>