… build/setEnvVariables.sh