| Question      | Answer
| ------------- | ------
| Bug fix?      | no
| New feature?  | no
| BC breaks?    | no
| Deprecations? | no
| Tests pass?   | yes
| Fixed tickets | 
| License       | MIT
| Doc PR        | 

### Description

This will reduce the time it taks for travis to run tests. 
We still run the same test suite but I've reduce the number of Travis builds. 

This PR will also reduce the risk of creating long build queues in Travis because we are only allowed 7? 8? concurrent builds. 

Before: 
 Elapsed time 17 min 33 sec
 Total time 1 hr 12 min 42 sec

Now: 
Elapsed time 5 min 58 sec
Total time 15 min 56 sec